@@ -3,9 +3,6 @@
   "version": "0.0.1",
   "description": "UI for the Rust playground",
   "main": "index.js",
-  "resolutions": {
-    "@types/react": "^17.0.2"
-  },
   "dependencies": {
     "@popperjs/core": "^2.4.0",
     "ace-builds": "^1.4.4",
@@ -14,18 +11,18 @@
     "history": "^5.3.0",
     "isomorphic-fetch": "^3.0.0",
     "lodash": "^4.17.0",
-    "monaco-editor": "^0.33.0",
+    "monaco-editor": "^0.34.1",
     "prismjs": "^1.6.0",
     "qs": "^6.4.0",
-    "react": "^17.0.1",
+    "react": "^18.2.0",
     "react-copy-to-clipboard": "^5.0.1",
-    "react-dom": "^17.0.1",
-    "react-monaco-editor": "^0.48.0",
+    "react-dom": "^18.2.0",
+    "react-monaco-editor": "^0.51.0",
     "react-popper": "^2.0.0",
     "react-portal": "^4.1.4",
     "react-prism": "^4.0.0",
     "react-redux": "^8.0.2",
-    "react-shadow": "^19.0.2",
+    "react-shadow": "^20.0.0",
     "redux": "^4.0.0",
     "redux-thunk": "^2.1.0",
     "regenerator-runtime": "^0.13.2",
@@ -43,19 +40,19 @@
     "@babel/preset-react": "^7.0.0",
     "@types/common-tags": "^1.8.1",
     "@types/isomorphic-fetch": "^0.0.36",
-    "@types/jest": "^27.0.0",
+    "@types/jest": "^29.2.5",
     "@types/lodash": "^4.14.118",
     "@types/prismjs": "^1.26.0",
     "@types/qs": "^6.9.7",
-    "@types/react": "^17.0.2",
+    "@types/react": "^18.0.26",
     "@types/react-copy-to-clipboard": "^5.0.4",
-    "@types/react-dom": "^17.0.1",
+    "@types/react-dom": "^18.0.10",
     "@types/react-portal": "^4.0.4",
     "@types/route-parser": "^0.1.4",
     "@typescript-eslint/eslint-plugin": "^5.9.0",
     "@typescript-eslint/parser": "^5.9.0",
     "autoprefixer": "^10.2.4",
-    "babel-loader": "^8.0.0",
+    "babel-loader": "^9.1.2",
     "babel-plugin-lodash": "^3.3.4",
     "basename": "^0.1.2",
     "compression-webpack-plugin": "^10.0.0",
@@ -66,7 +63,7 @@
     "eslint-plugin-react-hooks": "^4.0.0",
     "glob": "^8.0.3",
     "html-webpack-plugin": "^5.2.0",
-    "jest": "^28.1.0",
+    "jest": "^29.3.1",
     "json-loader": "^0.5.4",
     "mini-css-extract-plugin": "^2.0.0",
     "monaco-editor-webpack-plugin": "^7.0.1",
@@ -74,20 +71,20 @@
     "postcss": "^8.2.7",
     "postcss-loader": "^7.0.0",
     "postcss-nesting": "^10.1.1",
-    "postcss-simple-vars": "^6.0.3",
+    "postcss-simple-vars": "^7.0.1",
     "prettier": "^2.2.1",
     "stylelint": "^14.2.0",
     "stylelint-config-css-modules": "^4.0.0",
-    "stylelint-config-idiomatic-order": "^8.1.0",
+    "stylelint-config-idiomatic-order": "^9.0.0",
     "stylelint-config-prettier": "^9.0.3",
-    "stylelint-config-standard": "^25.0.0",
+    "stylelint-config-standard": "^29.0.0",
     "ts-essentials": "^9.0.0",
-    "ts-jest": "^28.0.3",
+    "ts-jest": "^29.0.4",
     "ts-loader": "^9.2.3",
     "typescript": "^4.2.2",
-    "typescript-plugin-css-modules": "^3.2.0",
+    "typescript-plugin-css-modules": "^4.1.1",
     "webpack": "^5.24.3",
-    "webpack-cli": "^4.5.0"
+    "webpack-cli": "^5.0.1"
   },
   "engines": {
     "node": "^16.13.1"
