@@ -24,7 +24,7 @@ const ChannelMenu: React.FC<ChannelMenuProps> = props => {
   const nightlyVersionDetails = useSelector(selectors.nightlyVersionDetailsText);
 
   const dispatch = useDispatch();
-  const changeChannel = useCallback((channel) => {
+  const changeChannel = useCallback((channel: Channel) => {
     dispatch(actions.changeChannel(channel));
     props.close();
   }, [dispatch, props]);
@@ -63,7 +63,7 @@ const ChannelMenu: React.FC<ChannelMenuProps> = props => {
   );
 };
 
-const Desc: React.FC<{}> = ({ children }) => (
+const Desc: React.FC<React.PropsWithChildren<unknown>> = ({ children }) => (
   <p className={styles.description}>{children}</p>
 );
 
