@@ -2,7 +2,7 @@ import React from 'react';
 
 import styles from './MenuAside.module.css';
 
-const MenuAside: React.FC = ({ children }) => (
+const MenuAside: React.FC<React.PropsWithChildren<unknown>> = ({ children }) => (
   <p className={styles.aside}>
     {children}
   </p>
